nile: Update builds to use docker
Update the nile repository to use docker for the kula builds. This was
done by updating the pyrex.ini file to use docker and set up the
appropriate variables. The create-build-nile.sh script was also created
to build the docker image, and nile-oe-init-build-env was created to
set up the build environment.

Signed-off-by: Erick Shepherd <erick.shepherd@ni.com>